Move from_unsigned_bytes code to SecretKey
